minor update 2

+++ arakelyan.daniil/P4/input_output_matrix.cpp
@@ -12,11 +12,11 @@ void arakelyan::inputMatrix(std::ifstream & input, int * matrix, size_t el)
     }
   }
 }
-std::ofstream & arakelyan::outputTransformMatrix(std::ofstream &output, double * smoothed, int rows, int cols)
+std::ofstream & arakelyan::outputTransformMatrix(std::ofstream &output, double * smoothed, size_t rows, size_t cols)
 {
-  for (int i = 0; i < rows; ++i)
+  for (size_t i = 0; i < rows; ++i)
   {
-    for (int j = 0; j < cols; ++j)
+    for (size_t j = 0; j < cols; ++j)
     {
       output << std::fixed << std::setprecision(1) << " " << smoothed[i * cols + j];
     }
